Moved EventBus class back to where it was in the original structure
