+++ buswear/src/main/java/pl/tajchert/buswear/wear/EventCatcher.java
@@ -3,6 +3,8 @@ package pl.tajchert.buswear.wear;
 import com.google.android.gms.wearable.MessageEvent;
 import com.google.android.gms.wearable.WearableListenerService;
 
+import pl.tajchert.buswear.EventBus;
+
 public class EventCatcher extends WearableListenerService {
 
     @Override
